memok6rt-order/src/main/java/com/sudipta/order/serviceImpl/OrderItemServiceImplementation.java
package com.sudipta.order.serviceImpl;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import com.sudipta.order.modal.OrderItem;
import com.sudipta.order.repository.OrderItemRepository;
import com.sudipta.order.service.OrderItemService;

@Service
public class OrderItemServiceImplementation implements OrderItemService {
	@Autowired
	private OrderItemRepository orderItemRepository;
	
	@Override
	public OrderItem createOrderItem(OrderItem orderItem) {
		
		return orderItemRepository.save(orderItem);
	}

}
